+++ src/gemseo_umdo/formulations/surrogate_settings.py
@@ -16,74 +16,18 @@
 
 from __future__ import annotations
 
-from collections.abc import Mapping  # noqa: TC003
-from collections.abc import Sequence  # noqa: TC003
-
-from gemseo.mlearning.regression.algos.base_regressor_settings import (
-    BaseRegressorSettings,  # noqa: TC002
+from gemseo_umdo.formulations.base_surrogate_settings import BaseSurrogate_Settings
+from gemseo_umdo.formulations.base_surrogate_settings import (
+    BaseSurrogateWithMCStatistics_Settings,
 )
-from gemseo.mlearning.regression.algos.rbf_settings import RBFRegressor_Settings
-from gemseo.utils.seeder import SEED
-from pydantic import Field
-from pydantic import NonNegativeInt
-from pydantic import PositiveInt
-
-from gemseo_umdo.formulations.base_sampling_settings import BaseSamplingSettings
+from gemseo_umdo.formulations.base_surrogate_settings import SurrogateQuality_Settings
 
 
-class Surrogate_Settings(BaseSamplingSettings):  # noqa: N801
+class Surrogate_Settings(  # noqa: N801
+    BaseSurrogate_Settings,
+    BaseSurrogateWithMCStatistics_Settings,
+    SurrogateQuality_Settings,
+):
     """The settings for the surrogate-based U-MDO formulation."""
 
     _TARGET_CLASS_NAME = "Surrogate"
-
-    regressor_settings: BaseRegressorSettings = Field(
-        default=RBFRegressor_Settings(), description="The regressor settings."
-    )
-
-    regressor_n_samples: PositiveInt = Field(
-        default=10000,
-        description="""The number of Monte Carlo samples
-to estimate the statistics from the regressor.""",
-    )
-
-    regressor_sampling_seed: NonNegativeInt = Field(
-        default=SEED, description="The seed of the Monte Carlo sampler."
-    )
-
-    quality_name: str = Field(
-        default="R2Measure",
-        description="The name of the measure to assess the quality of the regressor.",
-    )
-
-    quality_threshold: float | Mapping[str, float | Sequence[float]] = Field(
-        default=0.9,
-        description="The learning quality threshold below which a warning is logged.",
-    )
-
-    quality_cv_compute: bool = Field(
-        default=True,
-        description="Whether to estimate the quality by cross-validation (CV).",
-    )
-
-    quality_cv_n_folds: PositiveInt = Field(
-        default=5, description="The number of folds in the case of the CV technique."
-    )
-
-    quality_cv_randomize: bool = Field(
-        default=True,
-        description="""Whether to shuffle the samples
-before dividing them in folds in the case of the CV technique.""",
-    )
-
-    quality_cv_seed: NonNegativeInt | None = Field(
-        default=None,
-        description="""The seed of the pseudo-random number generator.
-
-If ``None``,
-an unpredictable generator is used.""",
-    )
-
-    quality_cv_threshold: float | Mapping[str, float | Sequence[float]] = Field(
-        default=0.8,
-        description="The CV quality threshold below which a warning is logged.",
-    )